requirements.txt
pyotp~=2.9.0
vrchatapi~=1.20.2
python-dotenv~=1.1.0
websockets~=11.0.3
cryptography~=45.0.6
aiomysql~=0.2.0
sqlmodel~=0.0.24
SQLAlchemy~=2.0.43